Refactor application store architecture
