+++ .changeset/few-coins-design.md
@@ -1,0 +1,5 @@
+---
+"wpgraphql-ide": major
+---
+
+Refactored stores, including renaming 'wpgraphql-ide' to 'wpgraphql-ide/app', and adding additional stores such as 'wpgraphql-ide/editor-toolbar
