+++ src/App.jsx
@@ -19,7 +19,7 @@ const setInitialState = () => {
 		setQuery,
 		setShouldRenderStandalone,
 		setInitialStateLoaded,
-	} = useDispatch( 'wpgraphql-ide' );
+	} = useDispatch( 'wpgraphql-ide/app' );
 
 	if ( isDedicatedIdePage ) {
 		setShouldRenderStandalone( true );
@@ -108,11 +108,11 @@ export function App() {
 
 export function RenderApp() {
 	const isInitialStateLoaded = useSelect( ( select ) => {
-		return select( 'wpgraphql-ide' ).isInitialStateLoaded();
+		return select( 'wpgraphql-ide/app' ).isInitialStateLoaded();
 	} );
 
 	const shouldRenderStandalone = useSelect( ( select ) => {
-		return select( 'wpgraphql-ide' ).shouldRenderStandalone();
+		return select( 'wpgraphql-ide/app' ).shouldRenderStandalone();
 	} );
 
 	if ( ! isInitialStateLoaded ) {
