+++ src/components/EditorDrawer.jsx
@@ -6,10 +6,10 @@ export function EditorDrawer( { children } ) {
 	const buttonLabel = '🚀';
 
 	const isDrawerOpen = useSelect( ( select ) => {
-		return select( 'wpgraphql-ide' ).isDrawerOpen();
+		return select( 'wpgraphql-ide/app' ).isDrawerOpen();
 	} );
 
-	const { setDrawerOpen } = useDispatch( 'wpgraphql-ide' );
+	const { setDrawerOpen } = useDispatch( 'wpgraphql-ide/app' );
 
 	return (
 		<div className="EditorDrawerRoot">
