+++ src/index.js
@@ -1,24 +1,23 @@
 // WordPress dependencies for hooks, data handling, and component rendering.
 import { createHooks } from '@wordpress/hooks';
-import { register } from '@wordpress/data';
 import { createRoot } from '@wordpress/element';
+import { registerStores, registerEditorToolbarButton } from './store';
 import * as GraphQL from "graphql/index.js";
 
 // Local imports including the store configuration and the main App component.
-import { store } from './store';
 import { App } from './App';
-
-// Register the store with wp.data to make it available throughout the plugin.
-register( store );
 
 // Create a central event hook system for the WPGraphQL IDE.
 export const hooks = createHooks();
 
+// Register all application stores.
+registerStores();
+
 // Expose a global variable for the IDE, facilitating extension through external scripts.
 window.WPGraphQLIDE = {
 	hooks,
-	store,
-	GraphQL
+	GraphQL,
+	registerEditorToolbarButton
 };
 
 /**
